get Article by Id works (Postman)

+++ back/src/main/java/com/openclassrooms/mddapi/controllers/ThemeController.java
@@ -19,7 +19,7 @@ public class ThemeController {
     }
 
     @GetMapping("/api/theme/{id}")
-    public Theme getThemeById(@PathVariable("id") Integer id){
+    public Theme getThemeById(@PathVariable("id") Long id){
         return themeService.findById(id);
     }
